fix_arm_warnings: Address comment

+++ database/sqlite/sqlite_aclk_chart.c
@@ -928,7 +928,7 @@ void aclk_update_retention(struct aclk_database_worker_config *wc)
             start_time = MIN(start_time, first_entry_t);
 
         if (memory_mode == RRD_MEMORY_MODE_DBENGINE && wc->chart_updates && (dimension_update_count < ACLK_MAX_DIMENSION_CLEANUP)) {
-            int live = ((now - last_entry_t) < (RRDSET_MINIMUM_DIM_LIVE_MULTIPLIER * update_every));
+            int live = ((uint32_t)(now - last_entry_t) < (RRDSET_MINIMUM_DIM_LIVE_MULTIPLIER * update_every));
             if (rc) {
                 first_entry_t = 0;
                 last_entry_t = 0;
